Use an environment called Web.Linux instead of Web.DEVCI.
This is because we need to use the same environment name for all
accounts when we're doing pipelines as code. This in turn is because we
have to specify the pipeline in the `.gocd.yaml` file in order to
register it with the environment, and because Go.CD doesn't give us the
option to have different `.gocd.yaml` files for different accounts.